0.5.8.3

+++ api/helpers.py
@@ -37,11 +37,10 @@ def createFolders():
 
 def handle_uploaded_files(zipRequest):
     withoutExt = str(zipRequest).split('.')[0]
-    write(zipRequest)
-    unzip(withoutExt, os.path.join("uploads/", str(zipRequest)))
+    # write(zipRequest)
+    unzip(withoutExt, zipRequest)
     validate = validate_package(withoutExt)
     if isinstance(validate, dict):
-        print('dict')
         moveIconsToStatic(withoutExt)
         new_json = reDefineJson(withoutExt)
         version = check_package_version(new_json)
@@ -92,17 +91,15 @@ def write(zip):
 
 def cleanup(packageName):
     from django.conf import settings
-    filesPath = os.path.join("files/", packageName)
-    uploadsPath = os.path.join(settings.BASE_DIR, "uploads/", packageName)
+    filesPath = os.path.join("files", packageName)
+    uploadsPath = os.path.join(settings.BASE_DIR, "uploads", packageName+'.zip')
 
     try:
-        log.new('Cleaning up '+packageName).logInfo()
         if os.path.exists(filesPath):
             rmtree(filesPath)
             os.remove(filesPath)
-
-        rmtree(uploadsPath)
-        os.remove(uploadsPath)
+        if os.path.exists(uploadsPath):
+            os.remove(uploadsPath)
     except Exception as e:
         log.new(e).logError()
         return False
